Refactor into components + css edits Divided components i App.js into different components, Home.js is now a "Container" and the other components are just stateless. Home should maybe be a class too, may edit.
