+++ webshop-redux/src/App.css
@@ -1,59 +1,36 @@
-.productList {
-	/* should use flexbox */
+
+#root, .productList{
+  display: flex;
+  justify-content: center;
+  text-align: center;
 }
+
 .productList > div:hover {
-	background-color: bisque;
-	border-color: black;
+  opacity: 0.7;
+  transition: opacity 200ms ease;
 }
-.productList > div {
-	background-color: beige;
-	border: 1px solid darkgray;
+.productList > div, .product {
+	background-color: #FFE5E5;
+	border: 1px solid rgba(200,200,200, 0.6);
 	border-radius: 4px;
-	margin: 12px;
-	padding: 8px;
+	margin: 1rem;
+	padding: 1rem;
 	display: inline-block;
 	min-height: 64px;
 	min-width: 64px;
 	text-align: center;
+  box-shadow: 0 0 20px rgba(200,200,200, 0.3);
+  transition: opacity 200ms ease;
+  cursor: pointer;
 }
 .productName {
 	font-weight: bold;
+  text-transform: capitalize;
 }
 .productPrice {
-	color: red;
-}
-h1 {
-	text-align: center;
-}
-.product {
-	background-color: aliceblue;
-	border: 1px solid darkgray;
-	border-radius: 4px;
-	text-align: center;
+	color: #7F7373;
 }
 
-/* ------------------------------------------------ */
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 80px;
-}
-
-.App-header {
-  background-color: #222;
-  height: 150px;
-  padding: 20px;
-  color: white;
-}
-
-.App-intro {
-  font-size: large;
-}
-
-@keyframes App-logo-spin {
-  from { transform: rotate(0deg); }
-  to { transform: rotate(360deg); }
+.product{
+  display: block;
 }
